Calculate read coverage bash script
SHARCNet bash submission script that takes a filtered VCF file (though filtering is not necessarily required for this to work) and calculates read coverage per individual per locus. In order to do this calls an R script outside of this script. Therefore this R script (calc_read_coverage.R) needs to be present in the the working directory.